+++ changelog.md
@@ -1,4 +1,9 @@
-﻿## 5.0.1
+﻿## 5.0.2
+- Debug: remove console.log for polls monitoring ([#202](https://github.com/Eccleria/ewibot/pull/202))
+- Feat: refacto on activities ([#204](https://github.com/Eccleria/ewibot/pull/204))
+- Fix: typo on anonymous check during poll refresh ([#199](https://github.com/Eccleria/ewibot/pull/199))
+
+## 5.0.1
 - Fix: botMessage missing setDefaultMemberPermission ([#184](https://github.com/Eccleria/ewibot/pull/184))
 - Fix: wrong commons usage from eventRoles buttons ([#190](https://github.com/Eccleria/ewibot/pull/190))
 - Fix: wrong arg order when using `removeAlavirien` ([#191](https://github.com/Eccleria/ewibot/pull/191))